Adding support for filtering a list of nodes by prefix.
versium.Versium:get_node_ids() now takes an optional table parameter.
If the table has a field "prefix", then only nodes that start with that
prefix are returned.  The table parameter is passed to the storage
implementation to make it possible for it to pre-filter the node.
However, at this point we do not count on the storage implementation to
do that.

Also changing sputnik-tickets to use the new method.