Added basic "admin_edit_ui" option.
The nodes now have two different fields for configuring edit UI: "admin_edit_ui" which
configures the interface shown to "Admin" and "edit_ui" which configures everyone else's
interface.  I used this separation as an opportunity to also remove most fields from the
non-Admin UI.  The only "advanced" options left are "Prototype" and "Category".

sputnik/actions/wiki.lua:actions.edit now takes an additional (fourth) parameter
edit_ui_field.  When this parameter is supplied, it's value is used, regardless of whether
the user is an admin or not.  If it is _not_ supplied, then either "admin_edit_ui" or
"edit_ui" is used depending on whether the user is "Admin".

_translations was updated with text for the label for the new field.