Adding support for permissions for "edit" actions.
The "edit" action in wiki.lua now checks if the user is allowed to edit the node, to keep them from
wasting time on edits that they won't be allowed to save.  @Lua_Config nodes are now blocked from editing
by non-Admins.

The fourth argument to actions.edit() is now a table called "etc".  All additional arguments should be
passed through this table.  In particular, what used to be the fourth parameter "edit_ui_field" is now
a etc.edit_ui_field.  A new field is added: etc.message_if_not_allowed, a message to be shown to the user
if they are not allowed to perform the edit action.

Note that actions.edit() doesn't check if the user can "edit" - it checks if the user can perform request.action.
This makes it possible for custom edit actions to set a few parameters and then let wiki.actions.edit do the work.
E.g.:

function actions.my_edit (node, request, sputnik)
   return wiki.actions.edit(node, request, sputnik, {edit_ui_field="my_edit_ui"})
end