Altered saci.sandbox to allow for a reset() function
This function when called, will clear the sandbox.values table of all of its elements.  This can be used to override inheritance up to a specific point (for example, you could do the following in an edit_ui field:

edit_ui = [=[reset(); foo = {1.30, "textarea"}]=]

As a result, all of the values that have already been set up until that point will be cleared, and foo will be the only field that is returned in edit_ui.

** THIS SHOULD NEVER BE USED IN AN ADMIN EDIT_UI FIELD, SINCE YOU WILL NEED TO EDIT THE PHYSICAL NODE IN ORDER TO RECOVER FROM IT **