Added code that allows for prototypes based on pattern matching.
Using the SPUTNIK_CONFIG.PROTOTYPE_PATTERNS table, specify a list of pattern/prototype pairs that are checked when a new node is being displayed.  This works by altering saci to return a second value from get_node which gets propogated through the entire system.  This return value is called "stub" and indicates that the node being returned is the stub of a non-existant node.

The first pattern that is matched in the table is applied to the node as it's prototype.