Adding a new mechanism for creating new nodes and updated sputnik-tickets.
Nodes now have a new field called "child_defaults".  The values of keys in child_default are used for as default
data for corresponding nodes.  E.g., "Tickets" has it's child_defaults set to:

    new = [[ prototype="@NewTicket"; title="New Ticket" ]]

This means that a request for Tickets/new will give us a node defined with [[ prototype="@NewTicket"; title="New
Ticket" ]].  Note that child_defaults field can be inherited, so it should be possible to do the same things as with
PROTOTYPE_PATTERNS.  E.g., if we want Tickets/Saci/new to create a new ticket associated with
"Tickets/saci", we just need to ensure that "Tickets/saci" has it's child_defaults set correctly.  However,
Tickets/saci can inherit this value of child_defaults, so we should only have to make sure that the prototype of
Tickets/Saci is set properly.