Adding support for metadata to sputnik/auth/simple.lua.
Note that I am adding support for "add_metadata" but not to "set_metadata" for now.
I want to treat set_metadata() as optional for now.

Also, fixing one place in mysql.lua where it called username:lower() without
capturing the output, and converting both modules to using functions in sputnik.auth.simple
for generating error messages.

Finally, changing the return values of "false" to "nil" in some places, to stick with
the convention that functions should either thrown an error or return _nil_ when something
went wrong.  Among other things, functions such as user_is_recent() should return "false"
if the user exist but is not recent, and "nil" if the user does not even exist.