Changed versium/saci to remove the need for get_node() to return metadata.
Versium's get_node() now just returns the data, without any revision
information.  This is done to make it possible to optimize
implementations for the case where we do not need the metadata.  For
this change to work implementation's get_node_info() now need to be able
to provide edit metadata for a specific revision.

This change introduce a complete separation between nodes' payload and
the versioning metadata.  get_node_info() and get_node_history() are now
the only ways to obtain such metadata.  The nodes themselves do not know
who their authors are, when they were edited, etc.  It is no longer
possible to do things like node.author or node.version.

As an additional change, I switched the parameter table for filedir to
expect the directory path as a numbered rather than named parameter.

The new Saci and Versium interfaces are documented at:

http://sputnik.freewisdom.org/en/Saci_API
http://sputnik.freewisdom.org/en/Versium