Integrating Jim's registration system with some changes to make it more configurable.
There are many changes, but here are the main ones:

1. recaptcha is offered as a separate rock.  it is ignored by default,
but can be turned on by setting CAPTCHA_MODULE and CAPTCHA_PARAMS
config values.

2. Anonymous users are allowed to edit by default, but must enter
captcha if CAPTCHA_MODULE is set.  Authenticated users do not need to do
captcha.

3. Email confirmation is activated if REQUIRE_EMAIL_CONFIRMATION is set.
In this case, registering a new user does not actually create an account
but rather creates a "ticket" an creating an account and emails the user
a link to activate the account.  If REQUIRE_EMAIL_CONFIRMATION is not
set then the user account is created immediately upon registration.

Captcha and email support will require a bit more testing.