Moving permission logic into saci.  Now, every Saci node has a method check_permissions(user, action) which returns true or false. Permissions rules are pulled from the "permissions" field and work the same way as before.
Note that saci at this point does not make any assumptions of how users
and actions are represented in the system and treats them as just IDs.

All Sputnik's permissions and groups are now supported as built-in
groups in Saci, _except_ "Admin", which is added on in Sputnik:init(),
incidentally providing an example of how to add groups to Saci
permissions.