Fix an issue with data serialiation
The previous version would fail if a string containing a newline also
ended with a closing square bracket (]).  For example the data:

I like to eat the following:
  * [Bananas]

Would have been serialized as:

[[I like to eat the following:
  * [Bananas]]]

Which is not valid Lua (as the final ] does not match anything
previously.  This commit resolves this bug and continues to work for all
other tested cases.