Fix default values in registration workflow
Previously if you incorrectly submitted the captcha or failed to agree
to the terms of service, the form you were presented with had all blank
fields. This change uses the request parameters for default form values
and clears the password and confirm password fields (no need to transmit
those more than necessary).